refactor(controller): move IntegrationClaims from token to controller
